+++ Tests/UZKArchiveTestCase.m
@@ -55,7 +55,8 @@ BOOL isAtLeast10_13SDK;
                            @"Test File A.txt",
                            @"Test File B.jpg",
                            @"Test File C.m4a",
-                           @"NotAZip-PK-ContentsUnknown"];
+                           @"NotAZip-PK-ContentsUnknown",
+                           @"Modified CRC Archive.zip"];
     
     NSArray *unicodeFiles = @[@"Ⓣest Ⓐrchive.zip",
                               @"Test File Ⓐ.txt",